Improve log when secret is not found in keyVault (#173)
This pull request introduces improvements to error handling and logging
when interacting with Azure Key Vault. The changes primarily focus on
providing more detailed error messages by including the path of the
secret that caused the error.

### Improvements to error handling:

*
[`src/Confix.Tool/src/Confix.Library/ThrowHelper.cs`](diffhunk://#diff-de2d44cf88807e7118a43465a75745e67e659f7268cad9c28496735aac87ad67L44-R48):
Modified the `SecretNotFound` method to accept an optional `path`
parameter and include it in the error message if provided.

### Updates to method signatures:

*
[`src/Confix.Tool/src/Confix.Library/Utilities/Azure/KeyVaultExtension.cs`](diffhunk://#diff-520f5459080dfd82d45c6e5bf94a9cc9a0ecda7e85b403aa953d50421aee0ebcL4-R17):
Updated the `HandleKeyVaultException` method to accept an optional
`path` parameter and pass it to the `SecretNotFound` method.

### Enhancements to Azure Key Vault provider:

*
[`src/Confix.Tool/src/Confix.Library/Variables/Providers/AzureKeyVault/AzureKeyVaultProvider.cs`](diffhunk://#diff-bcce2fb5047575d77b3d9a54ee0f9df0688c86e01fb6afd1665a8ae18b20f135L57-R57):
Updated the `ResolveAsync` and `SetAsync` methods to pass the `path`
parameter to the `HandleKeyVaultException` method.
[[1]](diffhunk://#diff-bcce2fb5047575d77b3d9a54ee0f9df0688c86e01fb6afd1665a8ae18b20f135L57-R57)
[[2]](diffhunk://#diff-bcce2fb5047575d77b3d9a54ee0f9df0688c86e01fb6afd1665a8ae18b20f135L75-R75)-
mention secret path in error message when available